+++ Arma3ObjectBuilder/utilities/validator.py
@@ -13,6 +13,9 @@ class ValidatorResult():
         self.is_valid = is_valid
         self.comment = comment
     
+    def __bool__(self):
+        return self.is_valid
+    
     def set(self, is_valid = True, comment = ""):
         self.is_valid = is_valid
         self.comment = comment
@@ -39,13 +42,13 @@ class ValidatorComponent():
 
         for item in strict:
             result = item()
-            if not result.is_valid:
+            if not result:
                 return False
         
         if warns_errs:
             for item in optional:
                 result = item()
-                if not result.is_valid:
+                if not result:
                     return False
         
         return True
@@ -57,13 +60,13 @@ class ValidatorComponent():
 
         for item in strict:
             result = item()
-            if not result.is_valid:
+            if not result:
                 self.logger.step("ERROR: %s" % result.comment)
                 is_valid = False
         
         for item in optional:
             result = item()
-            if not result.is_valid:
+            if not result:
                 self.logger.step("WARNING: %s" % result.comment)
                 is_valid &= not warns_errs
         
@@ -213,7 +216,7 @@ class ValidatorLODGeometry(ValidatorComponentLOD):
 
     def is_triangulated(self):
         result =  super().is_triangulated()
-        if result.is_valid:
+        if not result:
             result.set(False, "mesh is not triangulated (convexity is not definite)")
         
         return result
@@ -371,6 +374,26 @@ class ValidatorLODShadow(ValidatorComponentLOD):
             self.no_materials
         )
         optional = info = tuple()
+
+        return strict, optional, info
+
+
+class ValidatorLODUnderground(ValidatorLODGeometry, ValidatorLODShadow):
+    """LOD - Underground (VBS)"""
+
+    def conditions(self):
+        strict = (
+            self.is_contiguous,
+            self.is_triangulated,
+            self.is_convex,
+            self.has_components,
+            self.is_sharp,
+            self.no_materials
+        )
+        optional = tuple()
+        info = (
+            self.farthest_point,
+        )
 
         return strict, optional, info
 
@@ -599,10 +622,11 @@ class Validator():
         self.components = {
             **dict.fromkeys(("4", "26", "27", "28"), [ValidatorLODShadow]),
             **dict.fromkeys(("9", "10", "13"), [ValidatorLODPointcloud]),
-            **dict.fromkeys(("7", "8", "14", "15", "16", "17", "19", "20", "21", "22", "23", "24", "31"), [ValidatorLODGeometrySubtype]),
+            **dict.fromkeys(("7", "8", "14", "15", "16", "17", "19", "20", "21", "22", "23", "24"), [ValidatorLODGeometrySubtype]),
             "6": [ValidatorLODGeometry],
             "11": [ValidatorLODRoadway],
-            "12": [ValidatorLODPaths]
+            "12": [ValidatorLODPaths],
+            "30": [ValidatorLODUnderground]
         }
 
     def validate_lod(self, obj, lod, lazy = False, warns_errs = True, relative_paths = False):
@@ -623,6 +647,7 @@ class Validator():
             is_valid &= item(obj, bm, self.logger, relative_paths).validate(lazy, warns_errs)
 
         bm.free()
+        self.logger.step("Validation %s" % ("PASSED" if is_valid else "FAILED"))
         self.logger.level_down()
         self.logger.step("Finished validation")
 
@@ -639,7 +664,8 @@ class Validator():
 
         for item in components:
             is_valid &= item(skeleton, self.logger).validate(lazy, False)
-
+        
+        self.logger.step("Validation %s" % ("PASSED" if is_valid else "FAILED"))
         self.logger.level_down()
         self.logger.step("Finished validation")